Squashed commit of the following:
commit 1667096b0b66dbcb06d440821094cb2666bf5297
Author: xDekann <oozog255@gmail.com>
Date:   Sat Apr 8 23:01:22 2023 +0200

    Task 4 - distance between two users on globe

commit f78d2895a1137a50884b3f911e17e99e0a09b0f4
Author: xDekann <oozog255@gmail.com>
Date:   Sat Apr 8 21:44:26 2023 +0200

    Task 3 - getting most expensive cart + user fname

commit d0c1fe518b0eb23d466643ed9a7f883c061dae63
Author: xDekann <oozog255@gmail.com>
Date:   Sat Apr 8 20:15:20 2023 +0200

    Task 2 - Group by category with prices

commit d04b277ce39f1ad67dd8a4ba0a5656ffbbaafc6e
Author: xDekann <oozog255@gmail.com>
Date:   Sat Apr 8 19:53:16 2023 +0200

    Data initialization - task 1

+++ fake-api/src/main/java/org/api/fake/entity/user/Geolocation.java
@@ -3,8 +3,6 @@ package org.api.fake.entity.user;
 import com.fasterxml.jackson.annotation.JsonAlias;
 import lombok.*;
 
-import java.math.BigDecimal;
-
 @Setter
 @Getter
 @ToString
@@ -13,7 +11,7 @@ import java.math.BigDecimal;
 @NoArgsConstructor
 public class Geolocation {
     @JsonAlias("lat")
-    private BigDecimal latitude;
+    private double latitude;
     @JsonAlias("long")
-    private BigDecimal longitude;
+    private double longitude;
 }
